feat: Make @sentry/browser more treeshakeable
For Electron
